Implement Batch Normalization
Batch norm is another magic tool, with it finally i don't need to care
about how to initialize weight, and the normalization of input to cowork
with weight initial values.
With batch norm, everything could just be casual~. The only thing need to
do is to set gamma and beta in batch norm to fit your activation function.

I choose 1,1 to make a postive most distribution of each layers' input to
gain the effective gradient at first.